Big Refactoring: deleted Controllers with implementation via JdbcTemplate , left Controllers with implementation via Hibernate
