+++ myBrikProject/src/main/java/by/mybrik/controllers/ProductTypeController.java
@@ -1,4 +1,4 @@
-package by.mybrik.controllers.newRevision;
+package by.mybrik.controllers;
 
 import by.mybrik.controllers.requests.productTypeRequests.ProductTypeCreate;
 import by.mybrik.controllers.requests.productTypeRequests.ProductTypeUpdate;
