+++ myBrikProject/src/main/java/by/mybrik/controllers/StandardOrderController.java
@@ -1,4 +1,4 @@
-package by.mybrik.controllers.newRevision;
+package by.mybrik.controllers;
 
 import by.mybrik.controllers.requests.standardOrderRequests.StandardOrderCreate;
 import by.mybrik.domain.entities.StandardOrder;
